refactor(types): move the map pair read into MapType
MapMaterializer held a second copy of the Map wire format (7-bit length
prefix, then alternating key and value) and could not reach ClickHouseType's
protected ClearDBNull from outside the class, so it open-coded that too.
Both copies now live in MapType next to Read, which is where AGENTS.md puts
the read and write paths, and which matches ArrayType.ReadTyped<T>.

ReadPair now calls ClearDBNull instead of repeating the DBNull test.
PocoReadExpressionFactory keeps its MapType special case: the key and value
types are known only at parse time, so MapType cannot declare a closed
ITypedReader<List<KeyValuePair<K,V>>>. Only the body moved.

The boxed KeyValuePairs path (pairFactory) is untouched.

No behavior change. 10443 tests pass on net9.0.
